requirements.txt
aenum==3.1.16
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
arrow==1.3.0
attrs==25.3.0
certifi==2025.6.15
charset-normalizer==3.4.2
coloredlogs==15.0.1
daily-python==0.17.0
dataclasses-json==0.6.7
deepgram-sdk==4.1.0
deprecation==2.1.0
distro==1.9.0
elevenlabs==1.59.0
flatbuffers==25.2.10
flexcache==0.3
flexparser==0.4
frozenlist==1.7.0
future==1.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
humanfriendly==10.0
idna==3.10
jiter==0.10.0
llvmlite==0.44.0
loguru==0.7.3
Markdown==3.8.1
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.5.0
mypy_extensions==1.1.0
numba==0.61.2
numpy==1.26.4
onnxruntime==1.20.1
openai==1.70.0
packaging==25.0
pillow==11.1.0
Pint==0.24.4
pipecat==0.3.0
pipecat-ai==0.0.63
platformdirs==4.3.8
propcache==0.3.2
protobuf==5.29.5
pydantic==2.10.6
pydantic_core==2.27.2
pyloudnorm==0.1.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
requests==2.32.3
resampy==0.4.3
scipy==1.15.3
six==1.17.0
sniffio==1.3.1
soxr==0.5.0.post1
sympy==1.14.0
tqdm==4.67.1
types-python-dateutil==2.9.0.20250516
typing-inspect==0.9.0
typing_extensions==4.14.0
urllib3==2.5.0
websockets==15.0.1
yarl==1.20.1
